Separate the text view function
Still under development.
