+++ EyePedia/app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
 
             android:configChanges="screenSize|orientation"
 
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.EyePedia.NoActionBar">
             <intent-filter>
@@ -31,7 +31,7 @@
         </activity>
 
         <activity
-            android:name=".SettingActivity"
+            android:name=".activity.SettingActivity"
             android:label="Settings"/>
     </application>
 </manifest>
